+++ Assets/Scripts/Characters/Decks/EnemyScript.cs
@@ -7,8 +7,11 @@ public class Enemy : Units
     protected override void Start()
     {
         speed = 1f;
-        totalHealth = 200;
-        damagePerAttack = 15;
+        attackRate = 1.5f;
+        totalHealth = 3 * multiplierTotalHp;
+        damagePerAttack = 4;
+        defense = 4;
+
         manaCost = 3;
         entityType = EntityType.Basic;
         base.Start();