+++ Assets/Scripts/Core/GameLoopManager.cs
@@ -2,16 +2,15 @@ using System.Collections;
 using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
-using UnityEngine.AI;
+using UnityEngine.UI;
 
 public class GameLoopManager : MonoBehaviour
 {
     public static GameLoopManager instance;
-    public GameObject popupEndGame;
-    public bool isGameRunning { get; set; } = false;
 
-    private List<GameObject> playerUnits = new();
-    private List<GameObject> enemyUnits = new();
+    [Header("Popup End Game")]
+    [SerializeField] GameObject popupEndGame;
+    [SerializeField] Button buttonNextOrRestart;
 
     [Header("Deck & Mana manager")]
     public CardUI selectedCard;
@@ -19,9 +18,19 @@ public class GameLoopManager : MonoBehaviour
     [SerializeField] GameObject settings;
     public DeckData selectedDeck;
 
+
+    public bool isGameRunning = false;
+    private List<GameObject> playerUnits = new();
+    private List<GameObject> enemyUnits = new();
+
     void Awake()
     {
         instance = this;
+    }
+
+    void Start()
+    {
+        LevelInformationFadeTextManager.instance.DisplayTextWithFade(GameManager.instance.currentLevelData.currentlevel.ToString(), "Facile");
     }
 
     void Update()
@@ -84,7 +93,9 @@ public class GameLoopManager : MonoBehaviour
     {
         isGameRunning = false;
 
-        if (GameManager.instance.currentLevelData.chaptersAfterGame != "" && GameManager.instance.currentLevelData != null) {
+        BugTracker.Info("End of the game, player win: '" + isPlayerVictorious + "'.");
+
+        if (isPlayerVictorious && GameManager.instance.currentLevelData.chaptersAfterGame != "" && GameManager.instance.currentLevelData != null) {
             GameManager.instance.nextStory = GameManager.instance.currentLevelData.chaptersAfterGame;
             LoadingScene.Instance.LoadStory();
             return;
@@ -93,7 +104,13 @@ public class GameLoopManager : MonoBehaviour
         popupEndGame.SetActive(true);
         if (isPlayerVictorious) {
             popupEndGame.GetComponentInChildren<TextMeshProUGUI>().text = "You win";
+
+            buttonNextOrRestart.GetComponentInChildren<TextMeshProUGUI>().text = "Next";
+            buttonNextOrRestart.onClick.AddListener(GameManager.instance.SetNextLevel);
+            return;
         }
+
+        buttonNextOrRestart.onClick.AddListener(RestartGame);
     }
 
     public void CheckVictory()
@@ -123,15 +140,16 @@ public class GameLoopManager : MonoBehaviour
 
     public void RestartGame()
     {
+        BugTracker.Info("Game is restarting...");
+
         popupEndGame.SetActive(false);
-        foreach (GameObject key in playerUnits) {
-            key.SetActive(true);
+
+        foreach (GameObject key in playerUnits)
             key.GetComponent<Units>().ResetUnit();
-        }
-        foreach (GameObject key in enemyUnits) {
-            key.SetActive(true);
+
+        foreach (GameObject key in enemyUnits)
             key.GetComponent<Units>().ResetUnit();
-        }
+
         StartOrStopCombat(false);
 
         ManaManager.instance.AddMana(3);
@@ -148,12 +166,14 @@ public class GameLoopManager : MonoBehaviour
         foreach (GameObject e in entities) {
             e.GetComponentInChildren<Units>().isGameRunning = enabled;
         }
+        BugTracker.Info("Start Game: '"+isGameRunning+"'.");
     }
 
     private IEnumerator DesableEntityAfterDeath(GameObject entity)
     {
         yield return new WaitForSeconds(1f);
-        entity.SetActive(false);
+        if (isGameRunning)
+            entity.SetActive(false);
     }
 
 
@@ -162,17 +182,14 @@ public class GameLoopManager : MonoBehaviour
     public void SetSelectedCard(CardUI card)
     {
         if (selectedCard != null && selectedCard != card)
-        {
             selectedCard.DeselectCard();
-        }
 
         selectedCard = card;
     }
 
     public void DeselectCard()
     {
-        if (selectedCard != null)
-        {
+        if (selectedCard != null) {
             selectedCard.DeselectCard();
             selectedCard = null;
         }