+++ Assets/Scripts/Core/LevelSelectionManager.cs
@@ -28,10 +28,6 @@ public class LevelSelectManager : MonoBehaviour
 
             Button button = btn.GetComponent<Button>();
             button.onClick.AddListener(ui.OnClick);
-
-            // if (i != 1) {
-            //     button.interactable = false; // Only one level for now
-            // }
         }
     }
 }